Using os.path.abspath instead of realpath and normalize_path throughout
The use of os.path.realpath was incorrect as symbolic links should not
be followed. Changes to the output path via param.normalize path
eliminated in favour of an explicit output_directory parameter for
Launchers. The launch method may now be used directly in simple cases
without decorators. The root_directory_name method of Launcher is
removed in favour of a more useful get_root_directory method.