Simplified and cleaned up the DynamicArgs base class
Jobs can now update the state of the Dynamic Args object by writing to
the output stream instead of needing to generate a new file. This is
easier to understand and simplifies the implementation.  The next()
method has been moved to the base class: the behaviour of a Dynamics
Args class is now defined by two methods only (_update_state
_initial_state). The 'schedule' concept has been removed: the penalty
due to increased complexity outweighs the advantages of retaining it.